moved around and added docstrings
