+++ python/semantic_kernel/connectors/ai/open_ai/responses/open_ai_chat_response.py
@@ -3,7 +3,6 @@
 from typing import AsyncGenerator, Dict, List, Optional, Union
 
 from openai import AsyncStream
-from openai.types import Completion
 from openai.types.chat import ChatCompletion
 from openai.types.chat.chat_completion_chunk import ChatCompletionChunk, Choice
 from pydantic import PrivateAttr
@@ -12,57 +11,19 @@ from semantic_kernel.connectors.ai.ai_response import AIResponse
 from semantic_kernel.connectors.ai.open_ai.models.chat.function_call import FunctionCall
 
 
-class OpenAITextResponse(AIResponse):
-    "A text completion response from OpenAI."
-
-    raw_response: Union[Completion, AsyncStream[Completion]]
-    _parsed_content: Optional[Dict[str]] = PrivateAttr(default_factory=dict)
-
-    @property
-    def content(self) -> Optional[str]:
-        """Get the content of the response.
-
-        Content can be None when no text response was given, check if there are tool calls instead.
-        """
-        if not isinstance(self.raw_response, Completion):
-            raise ValueError("content is not available for streaming responses, use stream_content instead.")
-        return self.raw_response.choices[0].text
-
-    @property
-    def all_content(self) -> List[Optional[str]]:
-        """Get the content of the response.
-
-        Some or all content might be None, check if there are tool calls instead.
-        """
-        if not isinstance(self.raw_response, Completion):
-            if self._parsed_content is not {}:
-                return list(self._parsed_content.values())
-            raise ValueError("all_content is not available for streaming responses, use stream_content instead.")
-        return [choice.text for choice in self.raw_response.choices]
-
-    async def parse_stream(self) -> AsyncGenerator[str, None]:
-        """Get the streaming content of the response."""
-        if isinstance(self.raw_response, Completion):
-            raise ValueError("streaming_content is not available for regular responses, use content instead.")
-        async for chunk in self.raw_response:
-            if len(chunk.choices) == 0:
-                continue
-            for choice in chunk.choices:
-                self.parse_choice(choice)
-            if chunk.choices[0].delta.text:
-                yield chunk.choices[0].delta.text
-
-    def parse_choice(self, choice: Choice) -> None:
-        """Parse a choice and store the text."""
-        if choice.delta.content is not None:
-            if choice.index in self._parsed_content:
-                self._parsed_content[choice.index] += choice.delta.text
-            else:
-                self._parsed_content[choice.index] = choice.delta.text
-
-
 class OpenAIChatResponse(AIResponse):
-    """A response from OpenAI."""
+    """A response from OpenAI.
+
+    For streaming responses, make sure to async loop through parse_stream before trying anything else.
+    Once that is done:
+    - content: get the content of first choice of the response.
+    - all_content: get the content of all choices of the response.
+    - function_call: get the function call of first choice of the response.
+    - all_function_calls: get the function call of all choices of the response.
+    - tool_calls: get the tool calls of first choice of the response.
+    - all_tool_calls: get the tool calls of all choices of the response.
+    - parse_stream: get the streaming content of the response.
+    """
 
     raw_response: Union[ChatCompletion, AsyncStream[ChatCompletionChunk]]
     _parsed_content: Dict[int, str] = PrivateAttr(default_factory=dict)
@@ -175,7 +136,9 @@ class OpenAIChatResponse(AIResponse):
     async def parse_stream(self) -> AsyncGenerator[str, None]:
         """Get the streaming content of the response.
 
-        Any function calls, tool calls and content get's parsed and stored in parsed_content, function_call and tool_calls dict, with the index of the choice as the key.
+        Any content, function calls and tool calls get parsed and stored in:
+        parsed_content, function_call and tool_calls dicts,
+        with the index of the choice as the key.
 
         The first choice content is yielded as a stream.
         """
@@ -229,52 +192,3 @@ class OpenAIChatResponse(AIResponse):
                 self._parsed_content[choice.index] += choice.delta.content
             else:
                 self._parsed_content[choice.index] = choice.delta.content
-
-
-class AzureOpenAIChatResponse(OpenAIChatResponse):
-    """A response from Azure OpenAI."""
-
-    _tool_message_content: Dict[int, str] = PrivateAttr(default_factory=dict)
-
-    def parse_choice(self, choice: Choice) -> None:
-        super().parse_choice(choice)
-        if choice.delta.model_extra and "context" in choice.delta.model_extra:
-            if choice.index in self._tool_message_content:
-                for extra_context in choice.delta.model_extra.get("context", {}).get("messages", {}):
-                    if extra_context["role"] == "tool":
-                        self._tool_message_content[choice.index] += extra_context.get("content", "")
-            else:
-                for extra_context in choice.delta.model_extra.get("context", {}).get("messages", {}):
-                    if extra_context["role"] == "tool":
-                        self._tool_message_content[choice.index] = extra_context.get("content", "")
-
-    @property
-    def tool_message(self) -> Optional[str]:
-        """Get the tool content of the response."""
-        if not isinstance(self.raw_response, ChatCompletion):
-            if self._tool_message_content is not None:
-                return self._tool_message_content[0]
-            raise ValueError("tool_content is not available for streaming responses, use stream_tool_content instead.")
-        if self.tool_message_content:
-            return self.tool_message_content
-        if self.raw_response.model_extra and "context" in self.raw_response.model_extra:
-            for m in self.raw_response.model_extra["context"].get("messages", {}):
-                if m["role"] == "tool":
-                    self.tool_message_content = m.get("content", None)
-                    return self.tool_message_content
-
-    @property
-    def all_tool_messages(self) -> List[Optional[str]]:
-        """Get the tool content of the response."""
-        if not isinstance(self.raw_response, ChatCompletion):
-            if self._tool_message_content is not None:
-                return list(self._tool_message_content.values())
-            raise ValueError("tool_content is not available for streaming responses, use stream_tool_content instead.")
-        if self.tool_message_content:
-            return [self.tool_message_content]
-        if self.raw_response.model_extra and "context" in self.raw_response.model_extra:
-            return [
-                m.get("content", None)
-                for m in self.raw_response.model_extra["context"].get("messages", {})
-                if m["role"] == "tool"
-            ]
