+++ python/semantic_kernel/connectors/ai/open_ai/services/open_ai_text_completion_base.py
@@ -45,13 +45,6 @@ class OpenAITextCompletionBase(TextCompletionClientBase, OpenAIHandler):
             settings.ai_model_id = self.ai_model_id
         response = await self._send_request(request_settings=settings)
         return response
-        # if isinstance(response, Completion):
-        #     if len(response.choices) == 1:
-        #         return response.choices[0].text
-        #     return [choice.text for choice in response.choices]
-        # if len(response.choices) == 1:
-        #     return response.choices[0].message.content
-        # return [choice.message.content for choice in response.choices]
 
     async def complete_stream(
         self,
@@ -81,30 +74,6 @@ class OpenAITextCompletionBase(TextCompletionClientBase, OpenAIHandler):
         settings.stream = True
         response = await self._send_request(request_settings=settings)
         return response
-        # async for partial in response:
-        #     if len(partial.choices) == 0:
-        #         continue
-
-        #     if settings.number_of_responses > 1:
-        #         completions = [""] * settings.number_of_responses
-        #         for choice in partial.choices:
-        #             if hasattr(choice, "delta") and hasattr(choice.delta, "content"):  # Chat completion
-        #                 completions[choice.index] = choice.delta.content
-        #             elif hasattr(choice, "text"):  # Text completion
-        #                 completions[choice.index] = choice.text
-        #         if any(completions):
-        #             yield completions
-        #     else:
-        #         if hasattr(partial.choices[0], "delta") and hasattr(
-        #             partial.choices[0].delta, "content"
-        #         ):  # Chat completion
-        #             content = partial.choices[0].delta.content
-        #             if content:
-        #                 yield content
-        #         elif hasattr(partial.choices[0], "text"):  # Text completion
-        #             text = partial.choices[0].text
-        #             if text.strip():  # Exclude empty or whitespace-only text
-        #                 yield text
 
     def get_request_settings_class(self) -> "AIRequestSettings":
         """Create a request settings object."""
